requirements.txt
python-telegram-bot==20.0
selenium==4.0.0
beautifulsoup4==4.10.0
